+++ src/GenomeTrack1D.h
@@ -18,7 +18,7 @@
 
 class GenomeTrack1D : public GenomeTrack {
 public:
-	enum Functions { AVG, MIN, MAX, NEAREST, STDDEV, SUM, MAX_POS, MIN_POS, EXISTS, SIZE, SAMPLE, SAMPLE_POS, FIRST, FIRST_POS, LAST, LAST_POS, NUM_FUNCS };
+	enum Functions { AVG, MIN, MAX, NEAREST, STDDEV, SUM, LSE, MAX_POS, MIN_POS, EXISTS, SIZE, SAMPLE, SAMPLE_POS, FIRST, FIRST_POS, LAST, LAST_POS, NUM_FUNCS };
 
 	virtual ~GenomeTrack1D() {}
 
@@ -38,6 +38,7 @@ public:
 	float last_nearest() const { return m_last_nearest; }
 	float last_stddev() const { return m_last_stddev; }
 	float last_sum() const { return m_last_sum; }
+	float last_lse() const { return m_last_lse; }
 	float last_quantile(double percentile);
 	float last_exists() const { return m_last_exists; }
 	float last_size() const { return m_last_size; }
@@ -66,6 +67,7 @@ protected:
 	float        m_last_nearest;
 	float        m_last_stddev;
 	float        m_last_sum;
+	float        m_last_lse;
 	float        m_last_exists;
 	float        m_last_size;
 	float        m_last_sample;